Rename files to follow convention and add gemspec.
