+++ Dockerfile
@@ -8,6 +8,7 @@ WORKDIR $INSTALL_PATH
 
 ADD *.gemspec $INSTALL_PATH
 ADD Gemfile* $INSTALL_PATH
+ADD lib/at_home_in_vitro_test_kit/version.rb $INSTALL_PATH/lib/at_home_in_vitro_test_kit/version.rb
 RUN gem install bundler
 # The below RUN line is commented out for development purposes, because any change to the 
 # required gems will break the dockerfile build process.
