+++ lib/at_home_in_vitro_test_kit/http_request_group/request_fhir_bundle_entry_validation.rb
@@ -1,18 +1,18 @@
 require_relative '../test_common/fhir_bundle_validation'
 
-module AtHomeTestKit
-    class ManualFhirBundleEntryValidation < Inferno::Test
-        include AtHomeTestKit::FhirBundleValidator
+module AtHomeInVitroTestKit
+    class RequestFhirBundleEntryValidation < Inferno::Test
+        include AtHomeInVitroTestKit::FhirBundleValidator
       title 'FHIR Bundle Entry Validation'
       description %(
         Validate that the FHIR Bundle entries conform to their relative At-Home In-Vitro Test Report IG specifications.
       )
-      id :manual_fhir_bundle_entry_validation
-      input :fhir_bundle
+      id :request_fhir_bundle_entry_validation
+
+      uses_request :bundle
 
       def fhir_resource
-        resource = FHIR::Json.from_json(fhir_bundle);
-        resource
+        FHIR::Json.from_json(request.request_body)
       end
 
       run do
