+++ lib/at_home_in_vitro_test_kit/manual_entry_group/manual_fhir_bundle_entry_validation.rb
@@ -1,18 +1,18 @@
 require_relative '../test_common/fhir_bundle_validation'
 
-module AtHomeTestKit
-    class RequestFhirBundleEntryValidation < Inferno::Test
-        include AtHomeTestKit::FhirBundleValidator
+module AtHomeInVitroTestKit
+    class ManualFhirBundleEntryValidation < Inferno::Test
+        include AtHomeInVitroTestKit::FhirBundleValidator
       title 'FHIR Bundle Entry Validation'
       description %(
         Validate that the FHIR Bundle entries conform to their relative At-Home In-Vitro Test Report IG specifications.
       )
-      id :request_fhir_bundle_entry_validation
-
-      uses_request :bundle
+      id :manual_fhir_bundle_entry_validation
+      input :fhir_bundle
 
       def fhir_resource
-        FHIR::Json.from_json(request.request_body)
+        resource = FHIR::Json.from_json(fhir_bundle);
+        resource
       end
 
       run do
