+++ lib/at_home_in_vitro_test_kit/test_common/fhir_bundle_validation.rb
@@ -1,5 +1,5 @@
 
-module AtHomeTestKit
+module AtHomeInVitroTestKit
     module FhirBundleValidator
         def validate_bundle(resource)
