+++ lib/at_home_in_vitro_test_kit/version.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+module AtHomeInVitroTestKit
+    VERSION = '0.9.0'
+end
